Auth sort-of working, in progress
