+++ native/package.json
@@ -27,8 +27,11 @@
     "date-fns": "^2.11.1",
     "expo": "~36.0.0",
     "expo-camera": "~8.0.0",
+    "expo-random": "~8.0.0",
+    "expo-secure-store": "~8.0.0",
     "graphql": "^15.0.0",
     "graphql-tag": "^2.10.3",
+    "jwt-decode": "^2.2.0",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-hooks-global-state": "^1.0.0",
@@ -38,8 +41,8 @@
     "react-native-safe-area-context": "0.6.0",
     "react-native-screens": "2.0.0-alpha.12",
     "react-native-svg": "9.13.3",
-    "react-native-uuid": "^1.4.9",
-    "react-native-web": "~0.11.7"
+    "react-native-web": "~0.11.7",
+    "reactn": "^2.2.7"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
