+++ native/src/components/screens/LoadingScreen.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import { StyleSheet, View, ActivityIndicator } from 'react-native'
+
+export const LoadingScreen = () => {
+  return (
+    <View style={styles.container}>
+      <ActivityIndicator size="large" color="#0000ff" />
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+})
